Fix popcnt build constraints
